requirements.txt
bcrypt==3.2.2
blinker==1.4
cffi==1.15.0
click==8.1.3
cryptography==37.0.2
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.0.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.1.0
psycopg2-binary==2.9.3
pycparser==2.21
python-dateutil==2.8.2
qrcode==7.3.1
six==1.16.0
SQLAlchemy==1.4.36
Werkzeug==2.1.2
WTForms==3.0.1
